fix warning

+++ libraries/physics/src/ObjectAction.cpp
@@ -277,7 +277,7 @@ quint64 ObjectAction::localTimeToServerTime(quint64 timeValue) const {
     }
 
     qint64 serverClockSkew = getEntityServerClockSkew();
-    if (serverClockSkew < 0 && timeValue <= -serverClockSkew) {
+    if (serverClockSkew < 0 && timeValue <= (quint64)(-serverClockSkew)) {
         return 1; // non-zero but long-expired value to avoid negative roll-over
     }
 
@@ -291,7 +291,7 @@ quint64 ObjectAction::serverTimeToLocalTime(quint64 timeValue) const {
     }
 
     qint64 serverClockSkew = getEntityServerClockSkew();
-    if (serverClockSkew > 0 && timeValue <= serverClockSkew) {
+    if (serverClockSkew > 0 && timeValue <= (quint64)serverClockSkew) {
         return 1; // non-zero but long-expired value to avoid negative roll-over
     }
